Add MetadataTester
